backend/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.17
langchain-google-genai==0.0.8
google-generativeai==0.3.1
